requirements.txt
asgiref==3.7.2
certifi==2024.7.4
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.3
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-extensions==3.2.3
django-js-asset==2.2.0
filelock==3.13.1
idna==3.7
pillow==10.4.0
platformdirs==4.2.0
razorpay==1.4.2
requests==2.32.3
setuptools==72.2.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.25.1
